\name{pfclust-package}
\alias{pfclust-package}
\alias{pfclust}
\docType{package}
\title{
Logitudinal Profile cluster analysis
}
\description{
Package pfclust implements the PfCluster method
}
\details{
PFCluster method. More details are coming ...
}
\author{
Yaomin Xu
Maintainer: Yaomin Xu<yaomin@gmail.com>
}
